+++ tools/linter_lib/custom_check.py
@@ -292,6 +292,7 @@ python_rules = RuleList(
         {
             "pattern": "django.utils.translation",
             "include_only": {"test/", "zerver/views/development/"},
+            "exclude": {"zerver/views/development/dev_login.py"},
             "description": "Test strings should not be tagged for translation",
             "good_lines": [""],
             "bad_lines": ["django.utils.translation"],
